added enums to bytecode, fixed == and != operators not being defined in the compiler (how did i seriously miss this?) & fixed casts in bytecode completely disregarding the value to cast

+++ src/main/java/org/crayne/mu/bytecode/common/ByteCodeEnumMember.java
@@ -1,12 +1,12 @@
 package org.crayne.mu.bytecode.common;
 
-public record ByteCodeEnumMember(int enumId, long ordinal) {
+public record ByteCodeEnumMember(int enumId, int ordinal) {
 
     public int enumId() {
         return enumId;
     }
 
-    public long ordinal() {
+    public int ordinal() {
         return ordinal;
     }
